Chisel OpenAI sample to be more like existing samples
