+++ samples/openai/.gitignore
@@ -1,3 +1,2 @@
 .env
 node_modules
-out
