+++ samples/openai/package.json
@@ -1,23 +1,22 @@
 {
-  "name": "direqt-openai-example",
-  "version": "1.0.0",
-  "description": "Simple bot demonstrating how to use Direqt Messaging API with OpenAI's API",
-  "main": "index.js",
-  "scripts": {
-    "build": "tsc -p .",
-    "start": "npm run build && node ./out/index.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "direqt": "file:../..",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "openai": "^3.3.0"
-  },
-  "devDependencies": {
-    "typescript": "^5.1.3",
-    "@types/express": "^4.17.17"
-  }
+    "name": "direqt-openai-example",
+    "version": "1.0.0",
+    "description": "Simple bot demonstrating how to use Direqt Messaging API with OpenAI's API",
+    "main": "index.js",
+    "scripts": {
+        "start": "ts-node index.ts"
+    },
+    "author": "Direqt Inc.",
+    "license": "ISC",
+    "dependencies": {
+        "direqt": "file:../..",
+        "dotenv": "^16.3.1",
+        "express": "^4.18.2",
+        "openai": "^3.3.0"
+    },
+    "devDependencies": {
+        "@types/express": "^4.17.17",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.1.3"
+    }
 }
