--- samples/openai/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2016",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
-    "module": "commonjs",                                /* Specify what module code is generated. */
-    "sourceMap": true,                                /* Create source map files for emitted JavaScript files. */
-    "outDir": "./out",                                   /* Specify an output folder for all emitted files. */
-    "strict": true,                                      /* Enable all strict type-checking options. */
-  }
-}
